+++ src/libraries/System.Private.CoreLib/src/System/Globalization/TextInfo.WebAssembly.cs
@@ -14,7 +14,7 @@ namespace System.Globalization
             Debug.Assert(!GlobalizationMode.UseNls);
             Debug.Assert(GlobalizationMode.Hybrid);
 
-            int exceptionPtr = HasEmptyCultureName ?
+            nint exceptionPtr = HasEmptyCultureName ?
                 Interop.JsGlobalization.ChangeCaseInvariant(src, srcLen, dstBuffer, dstBufferCapacity, toUpper) :
                 Interop.JsGlobalization.ChangeCase(_cultureName, src, srcLen, dstBuffer, dstBufferCapacity, toUpper);
             if (exceptionPtr != IntPtr.Zero)
